+++ src/parse/ListObject.ts
@@ -1,18 +1,25 @@
-import { parseWarning, parsed } from "./common";
+import { nestWarningsUnderKey, type Parser } from "./common";
 import * as IR from "../IntermediateResult";
+import { evolve } from "../upstream/rambda";
 
-import type { Parser } from "./common";
+export const ListObject: Parser["ListObject"] = async function ({
+  element,
+  variable,
+  ctx,
+}) {
+  const parsedList = await this.ListArray({
+    element: element["@list"],
+    variable,
+    ctx,
+  });
 
-export const ListObject: Parser["ListObject"] = ({ element, variable }) =>
-  Promise.resolve(
-    parsed({
-      intermediateResult: new IR.LiteralValue(element),
-      term: variable,
-      warnings: [
-        parseWarning({
-          message:
-            "List objects are not yet supported. (https://github.com/m-ld/quild/issues/19)",
-        }),
-      ],
-    })
+  const parsedListObject = evolve(
+    {
+      intermediateResult: (ir) => new IR.Object({ "@list": ir }),
+      warnings: nestWarningsUnderKey("@list"),
+    },
+    parsedList
   );
+
+  return parsedListObject;
+};
