+++ src/parse/common.ts
@@ -171,6 +171,7 @@ export type Parse<
 export interface Parser {
   readonly Document: Parse;
   readonly NodeObjectArray: Parse<JsonArray, IR.Set>;
+  readonly ListArray: Parse<JsonArray>;
   readonly TopLevelGraphContainer: Parse<TopLevelGraphContainer, IR.Object>;
   readonly NodeObject: Parse<JsonObject, IR.Object>;
   readonly GraphObject: Parse;
